Ajout de la recuperation auto des data par cookies

+++ createams-back/src/main/java/com/myproject/backcreateams/repositories/UserRepository.java
@@ -6,7 +6,7 @@ import com.myproject.backcreateams.models.UserEntity;
 
 @Repository
 public interface UserRepository extends JpaRepository<UserEntity, Long> {
-    UserEntity findById(int id);
+    UserEntity findById(Long id);
     UserEntity findByMail(String mail);
     
 }
